+++ src/structures/vroom/input/input.h
@@ -65,6 +65,11 @@ public:
   std::vector<Job> jobs;
   std::vector<Vehicle> vehicles;
 
+  // Store rank in jobs accessible from job/pickup/delivery id.
+  std::unordered_map<Id, Index> job_id_to_rank;
+  std::unordered_map<Id, Index> pickup_id_to_rank;
+  std::unordered_map<Id, Index> delivery_id_to_rank;
+
   Input(unsigned amount_size);
 
   void set_geometry(bool geometry);